# crest-annotation-openapi/openapi.json
{
 "openapi": "3.0.2",
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "paths": {
...
  "/import/filesystem": {
   "post": {
    "tags": [
     "import"
    ],
    "summary": "Import Filesystem",
    "operationId": "import_filesystem",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Path",
       "type": "string"
      },
      "name": "path",
      "in": "query"
     },
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Commit",
       "type": "boolean",
       "default": false
      },
      "name": "commit",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FilesystemImport"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/import/iiif/3": {
   "post": {
    "tags": [
     "import"
    ],
    "summary": "Import Iiif3",
    "operationId": "import_iiif3",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Url",
       "type": "string"
      },
      "name": "url",
      "in": "query"
     },
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Commit",
       "type": "boolean",
       "default": false
      },
      "name": "commit",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Iiif3Import"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/import/iiif/2": {
   "post": {
    "tags": [
     "import"
    ],
    "summary": "Import Iiif2",
    "operationId": "import_iiif2",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Url",
       "type": "string"
      },
      "name": "url",
      "in": "query"
     },
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Commit",
       "type": "boolean",
       "default": false
      },
      "name": "commit",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Iiif2Import"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/import/digital-heraldry": {
   "post": {
    "tags": [
     "import"
    ],
    "summary": "Import Digital Heraldry",
    "operationId": "import_digital_heraldry",
    "parameters": [
     {
      "required": true,
      "schema": {
       "title": "Endpoint",
       "type": "string"
      },
      "name": "endpoint",
      "in": "query"
     },
     {
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      },
      "name": "project_id",
      "in": "query"
     },
     {
      "required": false,
      "schema": {
       "title": "Commit",
       "type": "boolean",
       "default": false
      },
      "name": "commit",
      "in": "query"
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_digital_heraldry_import_digital_heraldry_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DigitalHeraldryImport"
        }
       }
      }
     },
     "404": {
      "description": "Not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Body_import_digital_heraldry_import_digital_heraldry_post": {
    "title": "Body_import_digital_heraldry_import_digital_heraldry_post",
    "required": [
     "query"
    ],
    "type": "object",
    "properties": {
     "query": {
      "title": "Query",
      "type": "string"
     }
    }
   },
...
   "DigitalHeraldryImport": {
    "title": "DigitalHeraldryImport",
    "required": [
     "objects",
     "added",
     "problems"
    ],
    "type": "object",
    "properties": {
     "objects": {
      "title": "Objects",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/DigitalHeraldryObject"
      }
     },
     "added": {
      "title": "Added",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/DigitalHeraldryObject"
      }
     },
     "problems": {
      "title": "Problems",
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    }
   },
   "DigitalHeraldryObject": {
    "title": "DigitalHeraldryObject",
    "required": [
     "object_uuid",
     "object_data"
    ],
    "type": "object",
    "properties": {
     "object_uuid": {
      "title": "Object Uuid",
      "type": "string"
     },
     "object_data": {
      "$ref": "#/components/schemas/DigitalHeraldryObjectData"
     }
    }
   },
   "DigitalHeraldryObjectData": {
    "title": "DigitalHeraldryObjectData",
    "required": [
     "manifest",
     "bindings"
    ],
    "type": "object",
    "properties": {
     "manifest": {
      "title": "Manifest",
      "type": "string"
     },
     "image": {
      "title": "Image",
      "type": "string"
     },
     "bindings": {
      "title": "Bindings",
      "type": "object",
      "additionalProperties": {
       "type": "string"
      }
     },
     "type": {
      "title": "Type",
      "type": "string",
      "default": "dh",
      "const": "dh"
     }
    },
    "description": "Specific data for objects imported from Digital Heraldry"
   },
   "FilesystemImport": {
    "title": "FilesystemImport",
    "required": [
     "objects",
     "added"
    ],
    "type": "object",
    "properties": {
     "objects": {
      "title": "Objects",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/FilesystemObject"
      }
     },
     "added": {
      "title": "Added",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/FilesystemObject"
      }
     }
    }
   },
   "FilesystemObject": {
    "title": "FilesystemObject",
    "required": [
     "object_uuid",
     "object_data"
    ],
    "type": "object",
    "properties": {
     "object_uuid": {
      "title": "Object Uuid",
      "type": "string"
     },
     "object_data": {
      "$ref": "#/components/schemas/FilesystemObjectData"
     }
    }
   },
   "FilesystemObjectData": {
    "title": "FilesystemObjectData",
    "required": [
     "path"
    ],
    "type": "object",
    "properties": {
     "path": {
      "title": "Path",
      "type": "string"
     },
     "type": {
      "title": "Type",
      "type": "string",
      "default": "fs",
      "const": "fs"
     }
    },
    "description": "Specific data for objects imported from file system"
   },
   "HTTPValidationError": {
    "title": "HTTPValidationError",
    "type": "object",
    "properties": {
     "detail": {
      "title": "Detail",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      }
     }
    }
   },
   "Id": {
    "title": "Id",
    "maxLength": 65536,
    "minLength": 1,
    "type": "string",
    "format": "uri"
   },
   "Iiif2Import": {
    "title": "Iiif2Import",
    "required": [
     "objects",
     "added",
     "problems"
    ],
    "type": "object",
    "properties": {
     "title": {
      "title": "Title",
      "type": "string"
     },
     "objects": {
      "title": "Objects",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Iiif2Object"
      }
     },
     "added": {
      "title": "Added",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Iiif2Object"
      }
     },
     "problems": {
      "title": "Problems",
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    }
   },
   "Iiif2Object": {
    "title": "Iiif2Object",
    "required": [
     "object_uuid",
     "object_data"
    ],
    "type": "object",
    "properties": {
     "object_uuid": {
      "title": "Object Uuid",
      "type": "string"
     },
     "object_data": {
      "$ref": "#/components/schemas/Iiif2ObjectData"
     }
    }
   },
   "Iiif2ObjectData": {
    "title": "Iiif2ObjectData",
    "required": [
     "manifest",
     "sequence",
     "canvas",
     "service"
    ],
    "type": "object",
    "properties": {
     "manifest": {
      "title": "Manifest",
      "type": "string"
     },
     "sequence": {
      "title": "Sequence",
      "type": "string"
     },
     "canvas": {
      "title": "Canvas",
      "type": "string"
     },
     "image": {
      "title": "Image",
      "type": "string"
     },
     "service": {
      "$ref": "#/components/schemas/app__api__import_objects_iiif2__schemas__Service"
     },
     "type": {
      "title": "Type",
      "type": "string",
      "default": "iiif2",
      "const": "iiif2"
     }
    },
    "description": "Specific data for objects imported from IIIF 2.x"
   },
   "Iiif3Import": {
    "title": "Iiif3Import",
    "required": [
     "objects",
     "added",
     "problems"
    ],
    "type": "object",
    "properties": {
     "title": {
      "title": "Title",
      "type": "object",
      "additionalProperties": {
       "type": "array",
       "items": {
        "type": "string"
       }
      }
     },
     "display": {
      "title": "Display",
      "type": "string"
     },
     "objects": {
      "title": "Objects",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Iiif3Object"
      }
     },
     "added": {
      "title": "Added",
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Iiif3Object"
      }
     },
     "problems": {
      "title": "Problems",
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    }
   },
   "Iiif3Object": {
    "title": "Iiif3Object",
    "required": [
     "object_uuid",
     "object_data"
    ],
    "type": "object",
    "properties": {
     "object_uuid": {
      "title": "Object Uuid",
      "type": "string"
     },
     "object_data": {
      "$ref": "#/components/schemas/Iiif3ObjectData"
     }
    }
   },
   "Iiif3ObjectData": {
    "title": "Iiif3ObjectData",
    "required": [
     "manifest",
     "page",
     "annotation",
     "canvas",
     "service"
    ],
    "type": "object",
    "properties": {
     "manifest": {
      "title": "Manifest",
      "type": "string"
     },
     "page": {
      "title": "Page",
      "type": "string"
     },
     "annotation": {
      "title": "Annotation",
      "type": "string"
     },
     "canvas": {
      "title": "Canvas",
      "type": "string"
     },
     "service": {
      "$ref": "#/components/schemas/iiif_prezi3__skeleton__Service"
     },
     "type": {
      "title": "Type",
      "type": "string",
      "default": "iiif3",
      "const": "iiif3"
     }
    },
    "description": "Specific data for objects imported from IIIF 3.0"
   },
...
   "LngString": {
    "title": "LngString",
    "type": "object",
    "additionalProperties": false
   },
...
   "ServiceItem": {
    "title": "ServiceItem",
    "required": [
     "id",
     "type"
    ],
    "type": "object",
    "properties": {
     "id": {
      "$ref": "#/components/schemas/Id"
     },
     "type": {
      "title": "Type",
      "type": "string"
     },
     "label": {
      "$ref": "#/components/schemas/LngString"
     },
     "profile": {
      "title": "Profile",
      "type": "string"
     },
     "service": {
      "$ref": "#/components/schemas/iiif_prezi3__skeleton__Service"
     }
    }
   },
   "ServiceItem1": {
    "title": "ServiceItem1",
    "required": [
     "@id",
     "@type"
    ],
    "type": "object",
    "properties": {
     "@id": {
      "$ref": "#/components/schemas/Id"
     },
     "@type": {
      "title": "@Type",
      "type": "string"
     },
     "profile": {
      "title": "Profile",
      "type": "string"
     },
     "service": {
      "$ref": "#/components/schemas/iiif_prezi3__skeleton__Service"
     }
    }
   },
...
   "ValidationError": {
    "title": "ValidationError",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "type": "object",
    "properties": {
     "loc": {
      "title": "Location",
      "type": "array",
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      }
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    }
   },
   "app__api__import_objects_iiif2__schemas__Service": {
    "title": "Service",
    "required": [
     "@id",
     "@context"
    ],
    "type": "object",
    "properties": {
     "@id": {
      "title": "@Id",
      "type": "string"
     },
     "@context": {
      "title": "@Context",
      "type": "string"
     },
     "profile": {
      "title": "Profile",
      "type": "string"
     }
    }
   },
   "iiif_prezi3__skeleton__Service": {
    "title": "Service",
    "type": "array",
    "items": {
     "anyOf": [
      {
       "$ref": "#/components/schemas/ServiceItem"
      },
      {
       "$ref": "#/components/schemas/ServiceItem1"
      }
     ]
    }
   }
  }
 }
}